% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisLogn-package.R
\docType{package}
\name{poisLogn-package}
\alias{poisLogn-package}
\title{poisLogn: Functions for the Poisson-Lognormal Distribution and associated Poisson-Lognormal Count Regression Estimation}
\description{
Functions for the PMF, CDF, Quantile, and Random Number Generation for the Poisson-Lognormal Distribution. Also defines a `poisLogn` family and estimates various Poisson-Lognormal regression models. Inttegration of the Poisson-Lognormal PMF is accomplished using Gauss-Hermite Quadrature with 100 nodes/weights using Rcpp for speed and accuracy.
}
\author{
\strong{Maintainer}: Wood Jonathan \email{jwood2@iastate.edu}

}
\keyword{internal}
