% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_poislogn.R
\name{poisLogn}
\alias{poisLogn}
\title{Family Function for Poisson-Lognormal Distribution}
\usage{
poisLogn(sigma = 1)
}
\arguments{
\item{sigma}{A numeric scalar specifying the fixed dispersion parameter
(\eqn{\sigma > 0}) of the underlying Lognormal distribution.}
}
\description{
Creates a `family` object for the Poisson-Lognormal distribution to be used 
in model fitting functions like `glm()`. The family uses a log-link and 
requires a fixed dispersion parameter `sigma`.
}
\details{
The Poisson-Lognormal distribution is a compound distribution that can model
over-dispersed count data. The mean-variance relationship is:
\deqn{Var(Y) = E[Y] + (e^{\sigma^2}-1) E[Y]^2}
where \eqn{\sigma} is the dispersion parameter of the underlying Lognormal
distribution. As \eqn{\sigma \to 0}, the distribution converges to a
standard Poisson, and the variance converges to the mean.

The family uses a log-link function, where the linear predictor \eqn{\eta}
relates to the mean \eqn{\mu + exp(\sigma^2/2) = E[Y]} as \eqn{\eta = \log(\mu)}.


Since `glm` can only estimate one dispersion parameter (typically \eqn{\phi} 
in \eqn{Var(Y) = \phi V(\mu)}), the Lognormal dispersion parameter `sigma` 
must be provided as a fixed value to this family function. To estimate 
`sigma`, one might fit the model over a range of `sigma` values and select 
the one that maximizes the log-likelihood or minimizes AIC.
}
