% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poislogn.R
\name{PoissonLognormal}
\alias{PoissonLognormal}
\alias{dpLnorm}
\alias{ppLnorm}
\alias{qpLnorm}
\alias{rpLnorm}
\title{Poisson-Lognormal Distribution}
\usage{
dpLnorm(x, mu = 1, sigma = 1, log = FALSE)

ppLnorm(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

qpLnorm(p, mu = 1, sigma = 1)

rpLnorm(n, mu = 1, sigma = 1)
}
\arguments{
\item{x}{A vector of quantiles (non-negative integers).}

\item{mu}{A vector of positive means of the distribution.}

\item{sigma}{A vector of positive dispersion parameters.}

\item{log, log.p}{Logical; if `TRUE`, probabilities are given/returned as log(p).}

\item{q}{A vector of quantiles.}

\item{lower.tail}{Logical; if `TRUE` (the default), probabilities are
\eqn{P[X \le x]}; otherwise, \eqn{P[X > x]}.}

\item{p}{A vector of probabilities.}

\item{n}{The number of random numbers to generate.}
}
\value{
`dpLnorm` gives the density (PMF), `ppLnorm` gives the distribution
function (CDF), `qpLnorm` gives the quantile function, and `rpLnorm`
generates random deviates.
}
\description{
These functions provide density, distribution function, quantile function, and
random number generation for the Poisson-Lognormal (PLogN) distribution.
}
\details{
The Poiosson-Lognromal distribution is defined as a Poisson distribution with
a Lognormal-distributed mean. It is useful for modeling count data that are 
overdispersed with heavy tails. The mathematical formulation is:
\deqn{P(y)=\int_{-\infty}^\infty \frac{e^{-e^{log(\mu)+\sigma\epsilon}}\left(e^{log(\mu)+\sigma\epsilon}\right)^y}{y!\sqrt{2\pi}}e^{-\frac{\epsilon^2}{2}}d\epsilon}

Where \eqn{\mu} is the mean of the Poisson distribution, \eqn{\sigma} is the
dispersion parameter of the underlying Lognormal distribution, and \eqn{\epsilon}
is a standard normal variable. 

The mean of the distribution is given by:
\deqn{E[Y] = \mu e^{\sigma^2/2}}

The variance is given by:
\deqn{Var(Y) = E[Y] + (e^{\sigma^2} - 1) E[Y]^2}

The R package `poilog` is used for the underlying Poisson-Lognormal density.
}
\examples{
dpLnorm(c(0, 1, 2), mu = 1.5, sigma = 0.8)
ppLnorm(5, mu = 1.5, sigma = 0.8)
qpLnorm(0.95, mu = 1.5, sigma = 0.8)
rpLnorm(10, mu = 1.5, sigma = 0.8)
}
